refactor layout context
