+++ src/layout/App/Layout.tsx
@@ -1,8 +1,7 @@
-import { SidebarInset, SidebarProvider, useSidebar } from "@/components/ui/sidebar";
-import { LayoutContext } from "@/context/LayoutContext";
+import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import React from "react";
 import { Outlet } from "react-router-dom";
-const AppSidebar = React.lazy(() => import("@/components/app/Sidebar/Sidebar"))
+import SideBarLazy from "./SideBarLazy";
 
 export default function AppLayout(): React.JSX.Element {
     return <SidebarProvider>
@@ -11,28 +10,4 @@ export default function AppLayout(): React.JSX.Element {
             <Outlet />
         </SidebarInset>
     </SidebarProvider>
-}
-const SideBarLazy = (): React.JSX.Element => {
-    const { open, openMobile } = useSidebar()
-    const { screen } = React.useContext(LayoutContext)
-    const [isOpen, setOpen] = React.useState<boolean>(false)
-    React.useEffect(() => {
-        if (screen == "mobile") {
-            setOpen(openMobile)
-        } else {
-            setOpen(open)
-        }
-    }, [open, openMobile, screen])
-    return (
-        <>
-            {
-                isOpen ?
-                    <React.Suspense fallback="">
-                        <AppSidebar />
-                    </React.Suspense> : null
-            }
-
-        </>
-
-    )
 }
